yt_timesheet/requirements.txt
requests>=2.32.0
pandas>=2.2.0
openpyxl>=3.1.2
python-dateutil>=2.9.0
python-dotenv>=1.0.1